Extract memcache-related functionality from Context into its own object.
