+++ ndb/autobatcher.py
@@ -1,5 +1,5 @@
 #
-# Copyright 2008 The ndb Authors. All Rights Reserved.
+# Copyright 2017 The ndb Authors. All Rights Reserved.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
